requirements.txt
absl-py==0.1.11
astor==0.6.2
backports.functools-lru-cache==1.5
backports.weakref==1.0.post1
bleach==1.5.0
click==6.7
cycler==0.10.0
enum34==1.1.6
funcsigs==1.0.2
futures==3.2.0
gast==0.2.0
grpcio==1.10.0
html5lib==0.9999999
kiwisolver==1.0.1
Markdown==2.6.11
matplotlib==2.2.0
mock==2.0.0
numpy==1.14.1
opencv-python==3.4.0.12
pandas==0.22.0
pbr==3.1.1
Pillow==5.0.0
protobuf==3.5.2
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2018.3
scikit-learn==0.19.1
scipy==1.0.0
six==1.11.0
sklearn==0.0
subprocess32==3.2.7
tensorboard==1.6.0
tensorflow==1.6.0
termcolor==1.1.0
tf-unet==0.1.1
Werkzeug==0.14.1
